chore: rename status attestation files to status assertion
