+++ src/credential/status/index.ts
@@ -1,10 +1,10 @@
 import { type StartFlow } from "./01-start-flow";
-import { statusAssertion, type StatusAssertion } from "./02-status-attestation";
+import { statusAssertion, type StatusAssertion } from "./02-status-assertion";
 import { evaluateIssuerTrust, type EvaluateIssuerTrust } from "../issuance";
 import {
   verifyAndParseStatusAssertion,
   type VerifyAndParseStatusAssertion,
-} from "./03-verify-and-parse-status-attestation";
+} from "./03-verify-and-parse-status-assertion";
 
 export { evaluateIssuerTrust, statusAssertion, verifyAndParseStatusAssertion };
 export type {
